infravoice-backend/requirements.txt
# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# Pydantic
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# Cache & Background Tasks
redis==5.0.1
celery==5.3.6

# AI & ML
openai-whisper==20231117
google-generativeai==0.3.2
torch==2.1.2
torchaudio==2.1.2

# Utilities
requests==2.31.0
python-dotenv==1.0.0
aiofiles==23.2.1

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0
